backend/requirements.txt
Django>=3.0,<4.0
psycopg2>=2.8
requests>=2.0
djangorestframework